feat: Consolidate on a single LangGraph orchestration (#540)
This PR removes LangChain tools and Vertex AI Function Calling
orchestration and consolidates on LangGraph.

It also flattens the directory structure and refactors to simplify some
parts of code for easier understanding.

> [!NOTE]
> The failure in the integration test is expected. This PR is part of a
series of changes, and the corresponding fix for this test is in a
subsequent PR.
> ### Reasoning
> We've intentionally split the work into smaller, focused PRs to make
the review process more manageable and efficient.
> ### Merge Plan
> All related PRs will be merged into the `toolbox-main` branch first.
We will ensure all tests are passing on `toolbox-main` before merging
the entire feature set into `main`.